Move in-file tests to tests/ to aid refactoring better (too high error count, tests re-factoring is a separate task to tackle later) [ci skip]
